Revamping the "Dirty" functionality for Editing Existing Users
Previously, the method used to determine whether or not changes had been made
while editing an existing user was manipulation of a "global" ( to
SectionExistingUsers ) variable, `settingsAreDirty`. This caused problems when
reverting a previously updated value to it's original value, after which the UI
still indicated that changes had been made.

To resolve this problem we have replaced the `settingsAreDirty` boolean with
a call to a new function `userSettingsDirty` in `inDirtyState`.
`userSettingsDirty` returns true based on whether any of the User Setting fields
has a value different from their original value.

To finish off the change, each location that set `settingsAreDirty` has been
replaced with a call to `updateSaveIndicator`. Which just shows or hides the
save indicator based on `inDirtyState`.