Removing `hiddenName` property
This property was making it so that we could not lookup this component based on
the information returned by `Ext.form.Form.getBasicForm().getFieldValues()`. It
returned this value as opposed to the `id` of the component.

Manual testing was conducted to ensure that the updating of users was not
affected.