Update User Creation to include Manual Override
User creation did not properly set the `sticky` property when a user was
assigned the Unknown user and **not** the Unknown organization. `sticky` is set
to true when a user has been assigned the `Unknown` user but not the `Unknown`
organization.

This should have been included when the `Manual Override` checkbox was added to
the `Edit Existing Users`