Removing the last aRole::factory usage
In this commit we remove the last usage of `aRole::factory` and
`XDUser::assumeActiveRole`. The key to this was refactoring two functions
`XDUser::getAllRoles` and `XDUser::_getFOrmalRoleName`. `XDUser::getAllRoles` to
return strings ( acl names: 'cd', 'cs', usr' etc. ) instead of `aRole`
instances. And `XDUser::getFormalRoleName` to accept a new parameter
`$pubDisplay` which controls whether the function will return the public acl
name or display value if `$role_abbrev` is not found.

Now, the `XDuser::getAllRoles` change was possible because the only function
called on the results of `XDUser::getAllRoles` was `aRole->getIdentifier()`
which corresponds to `Acl::getName()` ( 'cd', 'cs', 'usr' etc. ). But there was
one, fairly ridiculous, side affect of not instantiating `aRole` classes which
necessitated the addition of `Query::initData($realm)`. It may be best to
illustrate this with a series of steps:

When `XDUser::getAllRoles` returns an array of `aRole` objects:
- User requests a timeseries chart from Metric Explorer
- ... stuff happens ..
- `XDUser::getAllRoles` is called and returns an array of `aRole` objects.
- this in turn calls `XDUser::assumeActiveRole($role)`
- this instantiates the role class that corresponds to $role
- During the Role Class instantiation all of the query descriptors for the role
are retrieved and instantiated.
- During the instantiation of each `QueryDescripter` it's realm's `Aggregate`
class has it's `registerStatistics` / `registerGroupBys` called. These are the
functions that ensure each Realms `Query` object has it's underlying statistic /
group by data structure populated.
- a `HighChartTimeSeries2` class is instantiated / configured.
- It looks up the requested metric in the realms statistic data structure that
has already been populated, succeeds in finding it and the chart is displayed
with data as expected.

When `XDUser::getAllRoles` no longer calls `XDUser::assumeActiveRole` we get the
following:
- User requests a timeseries chart from Metric Explorer
- ... stuff happens ...
- `XDUser::getAllRoles` is called and returns an array of strings that represent
this user's set of Roles.
- No `aRole` objects are instantiated so the realms statistics / group by data
structures are not populated.
- a `HighChartTimeSeries2` class is instantiated / configured.
- It looks up the requested metric in the **Not Populated** statistics data
structure and return's no `series` data. Thus presenting an empty chart to the
user.

So, by adding `Query::initData($realm)` and calling it after retrieving this
`Query`s realm in the following functions:
- &get_group_by_name_to_instance
- &get_group_by_name_to_class_name
- &get_statistic_name_to_class_name

We ensure that when the other `public static` functions of `Query` are called
that the underlying group by / statistic data structures are populated as
expected.

n