Active, Primary, Most Privileged Role Consolidation (#772)
* Removing Active / Primary related Code

This commit deals with the removal of code that interacts with the XDUser
active|primary role in anyway. The criteria for code being removed was if it
calls active|primary role or is called with active|primary role and is no longer
used.

* Updating `mostPrivilegedRole` to Acls

`getMostPrivilegedRole` will now return the most privileged Acl ( to be renamed
in a later acl ).

* Removing unused code / end points

These functions / end points are not currently in use and are being removed
ahead of the rest of the active|primary role consolidation work.

* Removing the last aRole::factory usage

In this commit we remove the last usage of `aRole::factory` and
`XDUser::assumeActiveRole`. The key to this was refactoring two functions
`XDUser::getAllRoles` and `XDUser::_getFOrmalRoleName`. `XDUser::getAllRoles` to
return strings ( acl names: 'cd', 'cs', usr' etc. ) instead of `aRole`
instances. And `XDUser::getFormalRoleName` to accept a new parameter
`$pubDisplay` which controls whether the function will return the public acl
name or display value if `$role_abbrev` is not found.

Now, the `XDuser::getAllRoles` change was possible because the only function
called on the results of `XDUser::getAllRoles` was `aRole->getIdentifier()`
which corresponds to `Acl::getName()` ( 'cd', 'cs', 'usr' etc. ). But there was
one, fairly ridiculous, side affect of not instantiating `aRole` classes which
necessitated the addition of `Query::initData($realm)`. It may be best to
illustrate this with a series of steps:

When `XDUser::getAllRoles` returns an array of `aRole` objects:
- User requests a timeseries chart from Metric Explorer
- ... stuff happens ..
- `XDUser::getAllRoles` is called and returns an array of `aRole` objects.
- this in turn calls `XDUser::assumeActiveRole($role)`
- this instantiates the role class that corresponds to $role
- During the Role Class instantiation all of the query descriptors for the role
are retrieved and instantiated.
- During the instantiation of each `QueryDescripter` it's realm's `Aggregate`
class has it's `registerStatistics` / `registerGroupBys` called. These are the
functions that ensure each Realms `Query` object has it's underlying statistic /
group by data structure populated.
- a `HighChartTimeSeries2` class is instantiated / configured.
- It looks up the requested metric in the realms statistic data structure that
has already been populated, succeeds in finding it and the chart is displayed
with data as expected.

When `XDUser::getAllRoles` no longer calls `XDUser::assumeActiveRole` we get the
following:
- User requests a timeseries chart from Metric Explorer
- ... stuff happens ...
- `XDUser::getAllRoles` is called and returns an array of strings that represent
this user's set of Roles.
- No `aRole` objects are instantiated so the realms statistics / group by data
structures are not populated.
- a `HighChartTimeSeries2` class is instantiated / configured.
- It looks up the requested metric in the **Not Populated** statistics data
structure and return's no `series` data. Thus presenting an empty chart to the
user.

So, by adding `Query::initData($realm)` and calling it after retrieving this
`Query`s realm in the following functions:
- &get_group_by_name_to_instance
- &get_group_by_name_to_class_name
- &get_statistic_name_to_class_name

We ensure that when the other `public static` functions of `Query` are called
that the underlying group by / statistic data structures are populated as
expected.

n

* Minor update to `Query::initData`

Just split out the `registerStatistics` and `registerGroupBys` so that they are
called based on `$_stats_initialized` and `$_group_by_initialized` respectively.

* Updating

Just a small update to the way previous initialization is being checked for.

* Updates per testing on metrics-dev

While testing on metrics-dev I ran into a problem where my Metric Explorer
queries were stored not as <acl.name> => <queries> but <acl.display> =>
<queries>. This caused a problem when
`MetricExplorer::convertActiveRoleToGlobalFilters` was called w/ the acl display
value generating an error stating "Unable to find configuration information for
Center Director".

To make sure that installs that may still have old style user query data hanging
around don't break I've updated the get queries end point to check if we have an
acl name or display, and if we have a display then retrieve the associated name.
A new function was added to Acls.php to help facilitate this, `getAclByDisplay`.

I tested this on xdmod-dev and did not run into a problem with old-style metric
explorer queries there.

* Add support for campus champion parameters

`Parameters::getParameters` needed to have a case `institution` to support
Campus Champions users.

* Updates per @jpwhite4

* Removing unneeded code

We no longer need:
```php
<?php
if (!isset(self::$_group_by_name_to_instance[$realm])) {
    self::$_group_by_name_to_instance[$realm] = array();
}
```
As `initData` will make sure that data is present.

* Updates per @plessbd

* derrrrrrr

* Removing unneeded code

This line of code is no longer needed as we're only ever going to be accessing
the first element of the exploded array.

* Updates per code review by @smgallo