Fix change tracking while editing existing users (#765)
* Revamping the "Dirty" functionality for Editing Existing Users

Previously, the method used to determine whether or not changes had been made
while editing an existing user was manipulation of a "global" ( to
SectionExistingUsers ) variable, `settingsAreDirty`. This caused problems when
reverting a previously updated value to it's original value, after which the UI
still indicated that changes had been made.

To resolve this problem we have replaced the `settingsAreDirty` boolean with
a call to a new function `userSettingsDirty` in `inDirtyState`.
`userSettingsDirty` returns true based on whether any of the User Setting fields
has a value different from their original value.

To finish off the change, each location that set `settingsAreDirty` has been
replaced with a call to `updateSaveIndicator`. Which just shows or hides the
save indicator based on `inDirtyState`.

* Removing `hiddenName` property

This property was making it so that we could not lookup this component based on
the information returned by `Ext.form.Form.getBasicForm().getFieldValues()`. It
returned this value as opposed to the `id` of the component.

Manual testing was conducted to ensure that the updating of users was not
affected.

* Style Fixes

* Update User Creation to include Manual Override

User creation did not properly set the `sticky` property when a user was
assigned the Unknown user and **not** the Unknown organization. `sticky` is set
to true when a user has been assigned the `Unknown` user but not the `Unknown`
organization.

This should have been included when the `Manual Override` checkbox was added to
the `Edit Existing Users`

* Rendered requirement

We now only check a field for changes if it's been rendered.

* Updates per code review comments @jpwhite4

- create_user.php:
  - updated per @jpwhite4 comment
- SectionExistingUsers.js
  - Moved `originalValue` population to the `fetchUserDetails` function along
  side the `setValue` / `initializeWithValue` calls. This means that the
  equivalent section from `userSettingsDirty` could be removed.
  - Removed the `String` casting from `userSettingsDirty` as we're setting the
  `originalValue` to what `getValue` will be providing.
  - Added a section of code to the Save Changes button callback that resets the
  fields original / startValue to `field.getValue()` so that we can accurately
  track changes after a user has been updated.

* Quick Update

Removal of out of date comments and slight simplification of code ( removed a
variable that wasn't needed ).

* Updating `sticky` in create_user

Need to check if `sticky` is present before `filter_var`ing it.

* Simplifying mapped to / institution / manual override population

- Updated `cmbUserMapping` to require a value be set.
- This means that we can get rid of the ternaries SectionNewUser.js@549 and just
straight up `parseInt(field.getValue())`.
- Also,  `institution`/ `mappedUser` closer to where the variables are actually
used.

* Style fixes

* Updating cmbUserType's originalValue type

When a user selects a new value for `cmbUserType` the type of this value is
string. Explicitly casting this to an int meant that it could never get
'un-dirty'.

* Updated where self.updateSaveIndicator was defined

Previously this function was defined near the bottom of the class. This meant
that when it was bound to AclGrid's `selectionHandler` `undefined` was actually
bound. This meant that when a user clicked on / off an acl selection that the
dirty user message would not be displayed.

* Extracted code into a new function `resetUserSettings`

- Moved code that reset the original values for user settings into it's own
  function. Replaced code moved w/ call to the new function.
- Updated `fetchUserDetails` to call this function so that the user settings
  dirty state would be accurate when selecting a new user.
- Updated the `updateSaveIndicator` function so that it's not accessible outside
  of `SectionExistingUsers`.

Updating `userSettingsDirty` && `resetUserSettings` scope

- These two new functions didn't need to be defined as `self.` either.