Replacing `ReportTemplateACL` w/ `report_template_acls`
`ReportTemplateACL` tracked which Roles are allowed access to which Report
Templates. As we're removing references to the `Roles` table, this has been
replaced by `report_template_acls` which does the same thing but with acls.

This commit adds the new `report_template_acls` table to the list of tables
managed by `acl-config` / ETLv2. It also updates
`XDReportManager::enumerateReportTemplates` to use the new
`report_template_acls` table.

`report_template_acls` will be populated from existing values in
`ReportTemplateACL`.

*Note: The reason we don't have a fk to `ReportTemplates` in
`report_template_acls` is that it's a MyISAM table.*