Updating report_template_acls population to ETLv2 json
After a conversation w/ @jpwhite4 it was clear that, for the sanity of people
interacting with this system going forward, it would be helpful to, when
associating an acl w/ a report template, be able to use the acl name not the
acl_id.

To accomplish this I've removed the `populate_report_template_acls.sql` file and
replaced it with some ETLv2 Json / Database Ingestor actions. The process steps
are now:
  - during `acls-xdmod-management`:
    - `moddb.report_template_acls` table is created.
    - `report_template_acls_staging` table is created / truncated and then
      populated  w/ the contents of
    `/etl/etl_data.d/acls/report-template-acls-staging.json`.
  - during `acls-import`:
    - `report_template_acls` is truncated / populated from
    `report_template_acls_staging`, translating the `acl_name` into `acl_id`.