Adding tests for `Configuration::toAssocArray`
During testing it was found that `Shredder.php` was requiring the return type of
`toAssocArray` to be an array. This is unlike our other configuration file usage
but was missed during the initial switch over of `Config` -> `Configuration`.
This commit adds tests to ensure that the code put in place to account for this
use case works as expected.

The tests added check the following:
  - that when the base file is an array with one object **and** the new
    `force_array_return` option is provided, that the results of `toAssocArray`
    is an array w/ one object.
  - that when the base file is an array with one object **and** the new
    `force_array_return` option is not provided, that the results of
    `toAssocArray` is the one object.
  - that when the base file is an array w/ multiple objects **and** the new
    `force_array_return` option is provided, that the results of `toAssocArray`
    is an array w/ multiple objects ( not: array(array( object, object, object))
  - that when the base file is an array w/ multiple objects **and** the new
    `force_array_return` is not provided, that the results of `toAssocArray` is
    an array w/ multiple objects.