Adding sorting to local config file processing
Previously the order in which local config files were read / processed was
non-deterministic. In this case, returned in an order as determined by the
underlying file system. It is desirable that we be able to control the order in
which these files are processed so the code has been changed to first, read all
the files from the specified directory in fs order, saving them in an array.
Next, sort this array in a deterministic array ( we default to sorting the
entries as strings ). Last, process these files as they were previously.