Remove UserRoleParameters Usage (#760)
* Updating Center[Director|Staff]Role `getActiveCenter`

Updated the sql used to retrieve a users associated organization id to reference
the User table instead of `UserRoleParameters` which is being removed. This is
possible because we now associate a user directly with an organization in the
User table.

* Updates to reflect changes in `Center[Director|Staff]Role`

The reason pi / user had organization's of -1 was due to Center[Director|Staff] being the only roles that would provide an organization id. This meant that only users that were assigned Center[Director|Staff] would have a non-negative one value. Now that all users are directly associated with an organization pi / user's expected values needed to be updated.

* Removing Unused sql that references UserRoleParameters

This PR removes UserRoleParameters so these sql files are no longer needed.

* Removing all usage of `UserRoleParameters`

This table is no longer used and references to it are being removed.
`user_acl_group_by_parameters` has been used instead for a while now, this is
just cleaning up the remnants.

* Changes per @jpwhite4 Code Review

Per @jpwhite4's code review comments I've made the following changes:
- Simplified the ternary @ index.php:243
- Removed test artifacts that are no longer used.

I also found the following change to be necessary after pulling in changes from
xdmod8.1
- Updated the logic in `arrayFilterKeysRecursive` so that discarding keys not in
the keyList is the first thing that's done.