Updates to  to support Shredder (#817)
* Updates to  to support Shredder

The  class requires the results of  as an array of arrays
regardless of how many resources there are. This change adds a new option to
that lets instantiators specify if they want the results of  to be wrapped in an array
if it would not normally be.

* word order, who needs that

* Adding tests for `Configuration::toAssocArray`

During testing it was found that `Shredder.php` was requiring the return type of
`toAssocArray` to be an array. This is unlike our other configuration file usage
but was missed during the initial switch over of `Config` -> `Configuration`.
This commit adds tests to ensure that the code put in place to account for this
use case works as expected.

The tests added check the following:
  - that when the base file is an array with one object **and** the new
    `force_array_return` option is provided, that the results of `toAssocArray`
    is an array w/ one object.
  - that when the base file is an array with one object **and** the new
    `force_array_return` option is not provided, that the results of
    `toAssocArray` is the one object.
  - that when the base file is an array w/ multiple objects **and** the new
    `force_array_return` option is provided, that the results of `toAssocArray`
    is an array w/ multiple objects ( not: array(array( object, object, object))
  - that when the base file is an array w/ multiple objects **and** the new
    `force_array_return` is not provided, that the results of `toAssocArray` is
    an array w/ multiple objects.

* Updated to utilize the new `force_array_return` option

These two callers of `toAssocArray` need to have the new `force_array_return`
option added to ensure that they always have an array to work with.

* Newline-palooza

- Adding trailing newlines to test artifact files that needed them.
- Updated `testToAssocArray` to add a newline to generated test artifacts.
- pretty printed the expected test artifacts

* Cleaning up test artifacts

- Removing a duplicate test artifact `resources_force_array_return.json`
- Updated the input for this test so that instead of expecting
  `resources_force_array_return.json` it now expects `resources.json`