Less infinite loops
So now that we have a class that handles the merging + extends resolving (
`XdmodConfiguration` ) we no longer need to rely on `Roles::getConfig`.

Usage of this function was replaced in:
  - Tabs.php@58
  - summary3.php@68,75

The one bit of functionality we lose by doing this is that, when a section
cannot be found we no longer attempt to look it up in the `default` acl.

To account for this, it was necessary to add a `summary_charts` section to the
`pub` acl. Which in turn required that this data be added to the expected
output in `artifacts/xdmod/acls/output/roles.json`.