Less infinite loops (#840)
* Less infinite loops

So now that we have a class that handles the merging + extends resolving (
`XdmodConfiguration` ) we no longer need to rely on `Roles::getConfig`.

Usage of this function was replaced in:
  - Tabs.php@58
  - summary3.php@68,75

The one bit of functionality we lose by doing this is that, when a section
cannot be found we no longer attempt to look it up in the `default` acl.

To account for this, it was necessary to add a `summary_charts` section to the
`pub` acl. Which in turn required that this data be added to the expected
output in `artifacts/xdmod/acls/output/roles.json`.

* Removing unused `use` statements

* Reverting roles.json and accounting for that in summary3

So instead of adding additional data to the pub acl we can instead implement the
implicit `default` extension in summary3. This means we don't have to migrate
any configuration files, Hoooray!

This change is per a conversation w/ @jpwhite4 where he talked some sense into
me.

* Updating expected test data

Updated the expected test data so that `pub` no longer has `summary_charts`

* Updating XdmodConfiguration usage w/ `assocArrayFactory`

Just removing boilerplate and replacing it w/
`XdmodConfiguration::assocArrayFactory`.

* Simplifying

* derrr