Add env var to force regrn tst harnss to gen expctd
The regression test harness used to automatically generate
expected results if the tests did not match. This then marked
the test as skipped. This is only desired behaiour if the developer
wants to generate test data. In normal CI we want the test to fail if
the actual does not match the expected.