Add environment variable to force regression test harness to generate expected results. (#848)
The regression test harness used to automatically generate
expected results if the tests did not match. This then marked
the test as skipped. This is only desired behaiour if the developer
wants to generate test data. In normal CI we want the test to fail if
the actual does not match the expected.

Also add an extra echo to the runtests.sh script to mitigate a bug in the shippable console folding.

Also switch the order of the tests run so the test generation will function correctly (the algorithm relies on the public user being run first).