Update config read order (#818)
* Adding sorting to local config file processing

Previously the order in which local config files were read / processed was
non-deterministic. In this case, returned in an order as determined by the
underlying file system. It is desirable that we be able to control the order in
which these files are processed so the code has been changed to first, read all
the files from the specified directory in fs order, saving them in an array.
Next, sort this array in a deterministic array ( we default to sorting the
entries as strings ). Last, process these files as they were previously.

* Adding test for local config read order

Just adding a test to ensure that local config files are read / processed in
alphabetical order.

* Adding newlines

* Updating test logic

Updated the `testLocalConfigReadOrder` function to not use a string comparison
as this was causing problems when run on Travis. Instead it is now using an
object comparison.

* Pretty printing expected output

Updates per @plessbd, just making sure that the expected output is pretty
printed.

* Update classes/Configuration/Configuration.php

Updates per code review by @jpwhite4

Co-Authored-By: ryanrath <ryanrath@buffalo.edu>

* Updates per code review comment by @jpwhite4

removing local scoped variable as it's not needed.

* Added a new helper function `assocArrayFactory`

Switching to the new `XdmodConfiguration` class incurred a lot of boilerplate
overhead. This helper function + making local config directories implicit when
instantiating a new `Configuration` object means that we can remove basically
all of the boilerplate that was added previously.

**NOTE:** The instances where `rawstatistics.json` was being read were not
          working as expected due to there being a type mismatch between the
          base config file and its local config files. This PR updates the way
          in which the config data is being used to be in line with how the
          config files will be going forward. There will be separate PR's to
          update the local config files, specifically in the SUPREMM model.
  # Please enter the commit message for your changes. Lines starting

* Formatting fixes

* Updated `local_dir` to be optional

the `datawarehouse` test specified in `xdmod_config.json` didn't need a
`local_dir` so this was removed. To account for some entries having a local_dir
and others not I updated the config object construction to account for this.

* Missed refactor

`$configFile` no longer exists so we instead use `CONFIG_DIR` for building the
file name to use in case of an exception.

* Updating rawstatistics.realms to an array

* Missed

Forgot add the `['realms']` qualification to the code that populates
`$realmExists`.

* Updating to utilize the new `assocArrayFactory`

Now that we have this nice new helper function it makes sense to use it.

* Update classes/Rest/Controllers/WarehouseControllerProvider.php

per @jpwhite4

Co-Authored-By: ryanrath <ryanrath@buffalo.edu>

* Fixing Syntax

* Only process `extends` if not a local config file