Addressing problem found by @jtpalmer (#849)
The problem is that when these entries were only in `composer.json` the
classpath was determined prior to run time, aka. before any other modules are
installed. So, if a module has code installed to any of these paths they will
not be found at round time. By moving them back into `linker.json` we resolve
this problem.