Remove Roles Classes (#773)
* Removing Active / Primary related Code

This commit deals with the removal of code that interacts with the XDUser
active|primary role in anyway. The criteria for code being removed was if it
calls active|primary role or is called with active|primary role and is no longer
used.

* Removing Roles and associated Tests and artifacts

Roles are now no longer being used and as such are being removed.

* Removing cruft from pull/rebase

This function sneaked in as part of the pull / rebase to include the latest
changes from xdmod8.1. It's a re-definition of `getAbbrevId` which obviously
caused a problem.

* Removing Unused Role related Class

This class is no longer referenced and can be removed.