Remove `Roles` Usage (#801)
* Updates to reflect changes in `Center[Director|Staff]Role`

The reason pi / user had organization's of -1 was due to Center[Director|Staff] being the only roles that would provide an organization id. This meant that only users that were assigned Center[Director|Staff] would have a non-negative one value. Now that all users are directly associated with an organization pi / user's expected values needed to be updated.

* Changes per @jpwhite4 Code Review

Per @jpwhite4's code review comments I've made the following changes:
- Simplified the ternary @ index.php:243
- Removed test artifacts that are no longer used.

I also found the following change to be necessary after pulling in changes from
xdmod8.1
- Updated the logic in `arrayFilterKeysRecursive` so that discarding keys not in
the keyList is the first thing that's done.

* Removing `UserRoles` usage in `XDStatistics::getUserVisitStats`

- XDStatistics
  - Added function doc block detailing how the function currently works.
  - `getUserVisitStats` SQL Updates:
    - Updated to utilize `user_acls` as opposed to `UserRoles`.
    - Updated so that it reports the counts per user as opposed to doing it in
      PHP.
    - Updated so that the `role_list` is quoted as opposed to taking care of
      it in the caller. This was done as `getUserVisitStats` is only called from
      one place.
    - Updated to include the UserTypes.type value instead of translating it to
      hard coded values in the caller.
  - Removed the code responsible for counting the Session Data as it's being
    done in SQL now.
- controller.php
  - Removed the code responsible for quoting the `role_list` as this is being
    done in the SQL now.
  - Removed the code responsible for translating the Users.user_type int to a
    hard coded string value that corresponds to UserTypes.type.
- UserAdminTest
  - Updated the `testGetUserVisits` code to keep track of all entries that are
    not found for more useful debugging.
  - Re-jiggered the order of the getUserVisits* tests so that the
    `testGetUserVisitsIncrements` test occurs last and doesn't mess with the
    other two tests.
  - Updated the `getUserVisits` helper function so that it always returns an
    `int`. Previously it would return strings causing problems when using the
    strict ( `===` ) tests.

* Removing unused code that references `Roles`

* Removing Tests that utilized removed functions

* Replacing `ReportTemplateACL` w/ `report_template_acls`

`ReportTemplateACL` tracked which Roles are allowed access to which Report
Templates. As we're removing references to the `Roles` table, this has been
replaced by `report_template_acls` which does the same thing but with acls.

This commit adds the new `report_template_acls` table to the list of tables
managed by `acl-config` / ETLv2. It also updates
`XDReportManager::enumerateReportTemplates` to use the new
`report_template_acls` table.

`report_template_acls` will be populated from existing values in
`ReportTemplateACL`.

*Note: The reason we don't have a fk to `ReportTemplates` in
`report_template_acls` is that it's a MyISAM table.*

* Updating report_template_acls population to ETLv2 json

After a conversation w/ @jpwhite4 it was clear that, for the sanity of people
interacting with this system going forward, it would be helpful to, when
associating an acl w/ a report template, be able to use the acl name not the
acl_id.

To accomplish this I've removed the `populate_report_template_acls.sql` file and
replaced it with some ETLv2 Json / Database Ingestor actions. The process steps
are now:
  - during `acls-xdmod-management`:
    - `moddb.report_template_acls` table is created.
    - `report_template_acls_staging` table is created / truncated and then
      populated  w/ the contents of
    `/etl/etl_data.d/acls/report-template-acls-staging.json`.
  - during `acls-import`:
    - `report_template_acls` is truncated / populated from
    `report_template_acls_staging`, translating the `acl_name` into `acl_id`.

* Fixing rebase snaffus

For some reason, a duplicate set of test / provide functions was added to
`UserAdminTest`. Those have been removed.

also added a newline at the end of
`configuration/etl/etl_tables.d/acls/report_template_acls.json`.

* Removing unneeded contents from acl-imports

This endpoint is already defined in the `defaults` section at the top of the
file. It does not need to be re-defined here.